+++ src/Catalog/AttributeType.php
@@ -10,6 +10,7 @@ namespace Wizaplace\SDK\Catalog;
 use MyCLabs\Enum\Enum;
 
 /**
+ * @method static AttributeType FREE()
  * @method static AttributeType FREE_NUMBER()
  * @method static AttributeType FREE_TEXT()
  * @method static AttributeType FREE_DATE()
@@ -22,6 +23,7 @@ use MyCLabs\Enum\Enum;
  */
 final class AttributeType extends Enum
 {
+    private const FREE = 'FREE';
     private const FREE_NUMBER = 'FREE_NUMBER';
     private const FREE_TEXT = 'FREE_TEXT';
     private const FREE_DATE = 'FREE_DATE';
@@ -39,6 +41,8 @@ final class AttributeType extends Enum
     public static function createFromLegacyMapping(string $value)
     {
         switch ($value) {
+            case 'F':
+                return self::FREE();
             case 'O':
                 return self::FREE_NUMBER();
             case 'T':