+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### New features
 
 - Added `\Wizaplace\SDK\Catalog\SearchProductAttribute::getType`
+- Added `\Wizaplace\SDK\Catalog\AttributeType::FREE`
 
 ### Bugfixes
 </details>