+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### New features
 
+- Added `\Wizaplace\SDK\Catalog\ProductAttribute::getValueIds`
 - Added `\Wizaplace\SDK\Catalog\SearchProductAttribute::getType`
 - Added `\Wizaplace\SDK\Catalog\AttributeType::FREE`